fix(hm/terminal/zellij): resolve keybinding conflicts by switching to Alt
When I was initially setting up Zellij, I made a mistake, more
specifically, a layering mistake. Both Ghostty and Zellij were
"competing" for control of the `Ctrl` key on Linux. This obviously
created a cursed workflow where Ghostty would "consume" keypresses
_before they ever reached Zellij_

The most obvious side effect was the hijacking of the `Ctrl+c`, which
made it basically impossible to interrupt any running commands

So to account for this:
- The `Ctrl` key is left untouched
- Zellij's primary modifier (`modKey`) is changed to `Alt`