feat(hm/terminal/zellij): add `tab,move` mod, misc
For a while I was using kind of a cursed Zellij config, so I decided to
look more into Zellij and try to come up with binds that make more sense
to me and capitalize more on the features Zellij has.

One of the major reasons for switching to Zellij was to leverage its
window management.

So what I did is the following:

- `Tab Mode` (`MOD+Shift+t`)
- `Move Mode` (`MOD+Shift+m`)
- `Scroll Mode` (`MOD+b`)

I'm still looking into floating panes; for now, I'm only making it
possible to "pop" a pane between non-floating and floating